% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNPR-package.R
\docType{package}
\name{DNPR-package}
\alias{DNPR}
\alias{DNPR-package}
\title{DNPR data in R}
\description{
DNPR package
}
\details{
Bridges the gap between DNPR2 and DNPR3.
}
\author{
Ann-Sophie Buchardt \email{ann-sophie.buchardt@regionh.dk}
}
