% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{duration}
\alias{duration}
\title{Create Duration Variable in DNPR3}
\usage{
duration(
  data,
  unit = c("hours", "minutes", "days", "seconds"),
  output = c("data.frame", "RPolarsDataFrame", "RPolarsLazyFrame", "csv", "parquet")
)
}
\arguments{
\item{data}{Data set}

\item{unit}{Duration unit. Possible values are seconds, minutes, hours (default), or days.}

\item{output}{Character string for desired output format.}
}
\value{
Data set with added duration variable
}
\description{
Compute contact duration from contact details.
}
\details{
Make sure that data covers the desired time period, e.g., dato_start in specific time interval
Make sure that data contains the required variables
}
\examples{
# Input example data from package and return object of class \code{data.frame}
files <- list.files(system.file("extdata", package = "DNPR"), pattern = "\\\\.csv$", ignore.case = TRUE, full.names = TRUE)
data <- read.csv(files[1])
head(data)
newdata <- duration(data)
head(newdata)
# Return duration in minutes
newdata <- duration(data, unit = "minutes")
head(newdata)

# Input example data from package (.csv file) and return object of class \code{data.frame}
files <- list.files(system.file("extdata", package = "DNPR"), pattern = "\\\\.csv$", ignore.case = TRUE, full.names = TRUE)
newdata <- duration(files)
# Return duration in minutes
newdata <- duration(files, unit = "minutes")

# Input example data from package (.parquet file) and return object of class \code{RPolarsDataFrame}
files <- list.files(system.file("extdata", package = "DNPR"), pattern = "\\\\.parquet$", ignore.case = TRUE, full.names = TRUE)
newdata <- duration(files, output = "RPolarsDataFrame")


}
