% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_type.R
\name{patient.type}
\alias{patient.type}
\title{Create In/Out Patient Type Variable in DNPR3}
\usage{
patient.type(dataset, method = c("cluster", "hybrid"))
}
\arguments{
\item{dataset}{Data set including the following variables
\itemize{
\item Either one variable containing contact duration in hours named \code{duration_h} or all of the following: \code{dato_start}, \code{tidspunkt_start}, \code{dato_slut}, and \code{tidspunkt_slut}
\item Either \code{elective} containing a logical value indicating whether contact is elective or acute *or* \code{prioritet} as defined in DNPR3.
\item Either \code{overnight} containing a logical value indicating whether contact is overnight or \code{dato_start} and \code{dato_end} containing start and end date of contact.
\item If \code{method = "hybrid_dep"}: p_overnight: Name of variable for percentage of overnight contacts.
}}

\item{method}{Name of patient type algorithm to use for classification.
Possible values are \code{method = "cluster"} (default) or \code{method = "hybrid"}.}
}
\value{
Data set with added patient type variable
}
\description{
Learns patient type (inpatient/acute outpatient/elective outpatient)
from contact details.
}
\examples{
# example code
# Using example data from package
files <- list.files(system.file("extdata", package = "DNPR"),
    pattern = "\\\\.csv$", ignore.case = TRUE, full.names = TRUE)
data <- files[1]
newdata <- read.csv(data)
head(newdata)
newdata <- patient.type(newdata)
head(newdata)

}
